Added menu bar and corresponding elements

+++ Form1.cs
@@ -2,6 +2,8 @@ using System.Diagnostics;
 using System.Windows.Forms;
 using System;
 using QRCoder;
+using System.Net.Http;
+
 
 namespace CleverDecksLauncher
 {
@@ -11,7 +13,8 @@ namespace CleverDecksLauncher
         private Process nodeProcess;
         private bool isCleverDecksRunning = false;
         private string[] cl_args;
-        private string qrCodeText = "";
+        private List<string> qrCodeText = new List<string> { };
+        private int qrCodeIndex = 0;
         public Form1(string[] args)
         {
             cl_args = args;
@@ -22,6 +25,94 @@ namespace CleverDecksLauncher
             notifyIcon1.DoubleClick += notifyIcon1_DoubleClick;
             this.Load += MainForm_Load;
             qrBox.Visible = false;
+
+            nextAddrButton.Visible = false;
+            previousAddrButton.Visible = false;
+            urlTextBox.ReadOnly = true;
+            urlTextBox.Visible = false;
+            textBox_clickToClose.Visible = false;
+
+            nextAddrButton.Click += NextAddrButton_Click;
+            previousAddrButton.Click += PreviousAddrButton_Click;
+
+            aboutCleverDecksLauncherToolStripMenuItem.Click += AboutCleverDecksLauncherToolStripMenuItem_Click;
+            showQRCodeToolStripMenuItem.Click += ShowQRCodeMenuItem_Click;
+
+            minimizeToTrayToolStripMenuItem.Click += MinimizeToTrayToolStripMenuItem_Click;
+            exitToolStripMenuItem1.Click += QuitMenuItem_Click;
+
+            removeLockFileOnRestartToolStripMenuItem.Click += RemoveLockFileOnRestartToolStripMenuItem_Click;
+
+            AboutForm aboutForm = new AboutForm();
+            aboutForm.Hide();
+        }
+
+        private void RemoveLockFileOnRestartToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            // create a file in the current directory to indicate that the lock file should be removed
+            System.IO.File.Create("remove_lock_file_on_restart").Close();
+        }
+
+        private void MinimizeToTrayToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            this.WindowState = FormWindowState.Minimized;
+            this.ShowInTaskbar = false;
+        }
+
+        private void AboutCleverDecksLauncherToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            AboutForm aboutForm = new AboutForm();
+            aboutForm.Show();
+        }
+
+        private void NextAddrButton_Click(object sender, EventArgs e)
+        {
+            if (qrCodeText.Count == 0)
+            {
+                ShowErrorNotification("QR Code not available. Please try again in a moment.");
+                return;
+            }
+            if (qrCodeIndex == qrCodeText.Count - 1)
+            {
+                qrCodeIndex = 0;
+            }
+            else
+            {
+                qrCodeIndex++;
+            }
+            UpdateQrCode(qrCodeText[qrCodeIndex]);
+        }
+
+        private void PreviousAddrButton_Click(object sender, EventArgs e)
+        {
+            if (qrCodeText.Count == 0)
+            {
+                ShowErrorNotification("QR Code not available. Please try again in a moment.");
+                return;
+            }
+            if (qrCodeIndex == 0)
+            {
+                qrCodeIndex = qrCodeText.Count - 1;
+            }
+            else
+            {
+                qrCodeIndex--;
+            }
+            UpdateQrCode(qrCodeText[qrCodeIndex]);
+        }
+
+        private void UpdateQrCode(string qrData)
+        {
+            using (QRCodeGenerator qrGenerator = new QRCodeGenerator())
+            using (QRCodeData qrCodeData = qrGenerator.CreateQrCode(qrData, QRCodeGenerator.ECCLevel.Q))
+            using (QRCode qrCode = new QRCode(qrCodeData))
+            {
+                Bitmap qrCodeImage = qrCode.GetGraphic(8);
+                qrBox.Image = qrCodeImage;
+                qrBox.Visible = true;
+                urlTextBox.Visible = true;
+                urlTextBox.Text = qrData;
+            }
         }
 
         private void MainForm_FormClosing(object sender, FormClosingEventArgs e)
@@ -66,18 +157,31 @@ namespace CleverDecksLauncher
         private void qrBox_Click(object sender, EventArgs e)
         {
             qrBox.Visible = false;
+            previousAddrButton.Visible = false;
+            nextAddrButton.Visible = false;
+            urlTextBox.Visible = false;
+            textBox_clickToClose.Visible = false;
         }
 
         private void LaunchNodeApp()
         {
+            // check to see if the remove_lock_file_on_restart file exists
+            bool removeLockFile = false;
+            if (System.IO.File.Exists("remove_lock_file_on_restart"))
+            {
+                System.IO.File.Delete("remove_lock_file_on_restart");
+                System.IO.File.Delete("lock");
+                removeLockFile = true;
+            }
             var startInfo = new ProcessStartInfo
             {
                 CreateNoWindow = true,
                 UseShellExecute = false,
                 RedirectStandardOutput = true,
                 RedirectStandardInput = true,
+                RedirectStandardError = true,
             };
-            
+
             if (cl_args.Length == 0 && CheckFileExists("CleverDecks.exe"))
             {
                 startInfo.FileName = "CleverDecks.exe";
@@ -89,7 +193,11 @@ namespace CleverDecksLauncher
             else if (cl_args.Length == 1 && cl_args[0].EndsWith(".js") && CheckFileExists(cl_args[0]))
             {
                 startInfo.FileName = "node";
-                startInfo.Arguments = cl_args[0];
+                startInfo.ArgumentList.Add(cl_args[0]);
+                if (removeLockFile)
+                {
+                    startInfo.ArgumentList.Add("--remove_lock");
+                }
                 int lastSlash = cl_args[0].LastIndexOf('/');
                 if (lastSlash == -1)
                 {
@@ -100,35 +208,26 @@ namespace CleverDecksLauncher
             else
             {
                 bool fileExists = false;
-                if(cl_args.Length == 1)
+                if (cl_args.Length == 1)
                 {
                     fileExists = CheckFileExists(cl_args[0]);
                 }
-                ShowErrorNotification("Unable to launch CleverDecks." + (fileExists?"":" Specified file does not exist."));
+                ShowErrorNotification("Unable to launch CleverDecks." + (fileExists ? "" : " Specified file does not exist."));
                 Application.Exit();
             }
-            var startsInfo = new ProcessStartInfo
-            {
-                FileName = "node",
-                Arguments = "index.js",
-                WorkingDirectory = "O:/Projects/CleverDecks",
-                CreateNoWindow = true,
-                UseShellExecute = false,
-                RedirectStandardOutput = true,
-                RedirectStandardInput = true,
-            };
+
             try
             {
                 nodeProcess = Process.Start(startInfo);
             }
             catch (Exception ex)
             {
-                ShowErrorNotification("Unable to launch CleverDecks." + ex.Message.Substring(0,ex.Message.Length>64?64:ex.Message.Length));
+                ShowErrorNotification("Unable to launch CleverDecks." + ex.Message.Substring(0, ex.Message.Length > 64 ? 64 : ex.Message.Length));
                 Application.Exit();
             }
             nodeProcess.OutputDataReceived += NodeProcess_OutputDataReceived;
             nodeProcess.BeginOutputReadLine();
-            
+
             isCleverDecksRunning = true;
 
         }
@@ -156,22 +255,27 @@ namespace CleverDecksLauncher
                 // Invoke the UI thread to update the TextBox
                 this.Invoke(new Action(() =>
                 {
-                    if (e.Data.Contains("http://") && qrCodeText.Equals(""))
+                    if (e.Data.Contains("http://"))
                     {
                         // extract the http address from the log
                         int start = e.Data.IndexOf("http://");
-                        qrCodeText = e.Data.Substring(start);
+                        qrCodeText.Add(e.Data.Substring(start));
                     }
                     textBox1.AppendText(e.Data + Environment.NewLine);
-                    if(autoScrollCheckBox.Checked)
+                    if (autoScrollCheckBox.Checked)
                     {
                         textBox1.SelectionStart = textBox1.Text.Length;
                         textBox1.ScrollToCaret();
                     }
+                    if (e.Data.Contains("GET /exit") || e.Data.Contains("Goodbye!"))
+                    {
+                        nodeProcess.WaitForExit(10000);
+                        nodeProcess?.Kill();
+                        Application.Exit();
+                    }
                 }));
             }
         }
-
 
         private void ShowOutputMenuItem_Click(object sender, EventArgs e)
         {
@@ -202,15 +306,15 @@ namespace CleverDecksLauncher
 
         private void LaunchBrowserMenuItem_Click(object sender, EventArgs e)
         {
-            if(qrCodeText.Equals(""))
+            if (qrCodeText.Count == 0)
             {
-                ShowErrorNotification("QR Code not available. Please try again later.");
+                ShowErrorNotification("Browser launch not available. Try again in a moment.");
                 return;
             }
-            
+
             var startInfo = new ProcessStartInfo
             {
-                FileName = qrCodeText,
+                FileName = qrCodeText[qrCodeIndex],
                 UseShellExecute = true
             };
             try
@@ -223,37 +327,85 @@ namespace CleverDecksLauncher
             }
         }
 
-        private void QuitMenuItem_Click(object sender, EventArgs e)
+        private async void QuitMenuItem_Click(object sender, EventArgs e)
         {
-            var quitTask = Task.Run(() =>
+            if (!nodeProcess.HasExited)
             {
+                nodeProcess.StandardInput.AutoFlush = true;
                 nodeProcess.StandardInput.WriteLine("terminate");
+                var exitUrl = qrCodeText + "/exit";
+                // extract the port number from qrCodeText (http://localhost:3000/)
+                string url = qrCodeText.ElementAt(qrCodeIndex);
+                string port = url.Substring(url.LastIndexOf(":") + 1);
+                if (port.Contains("/"))
+                {
+                    port = port.Substring(0, port.IndexOf('/'));
+                }
+                ShowErrorNotification("Exit URL: " + exitUrl);
+                exitUrl = "http://localhost:" + port + "/exit";
+                // create http request to exit
+                try
+                {
+                    await new HttpClient().GetAsync(exitUrl);
+                }
+                catch (Exception ex)
+                {
+                    ShowErrorNotification("Failed to send exit request to CleverDecks.");
+                }
                 nodeProcess.WaitForExit(5000);
-            });
-
-            quitTask.Wait(5000);
-            nodeProcess?.Kill();
+            }
+            if (!nodeProcess.HasExited)
+            {
+                nodeProcess?.Kill();
+            }
+            nodeProcess.WaitForExit(5000);
             Application.Exit();
         }
 
         private void ShowQRCodeMenuItem_Click(object sender, EventArgs e)
         {
-            if (qrCodeText.Equals(""))
+            if (qrCodeText.Count == 0)
             {
-                ShowErrorNotification("QR Code not available. Please try again later.");
+                ShowErrorNotification("QR Code not available. Please try again in a moment.");
             }
             else
             {
-                using (QRCodeGenerator qrGenerator = new QRCodeGenerator())
-                using (QRCodeData qrCodeData = qrGenerator.CreateQrCode(qrCodeText, QRCodeGenerator.ECCLevel.Q))
-                using (QRCode qrCode = new QRCode(qrCodeData))
+                this.ShowInTaskbar = true;
+                previousAddrButton.Visible = true;
+                nextAddrButton.Visible = true;
+                textBox_clickToClose.Visible = true;
+                qrBox.Visible = true;
+                qrBox.BringToFront();
+                urlTextBox.Visible = true;
+                urlTextBox.BringToFront();
+                nextAddrButton.BringToFront();
+                previousAddrButton.BringToFront();
+                this.WindowState = FormWindowState.Normal;
+                UpdateQrCode(qrCodeText[qrCodeIndex]);
+            }
+        }
+
+        private void ShowQRCodeButton_Click(object sender, EventArgs e)
+        {
+            if (qrCodeText.Count == 0)
+            {
+                ShowErrorNotification("QR Code not available. Please try again in a moment.");
+            }
+            else
+            {
+                if (qrBox.Visible == true)
                 {
-                    Bitmap qrCodeImage = qrCode.GetGraphic(8);
-                    qrBox.Image = qrCodeImage;
-                    qrBox.Visible = true;
-                    this.ShowInTaskbar = true;
-                    this.WindowState = FormWindowState.Normal;
+                    qrBox.Visible = false;
+                    previousAddrButton.Visible = false;
+                    nextAddrButton.Visible = false;
+                    urlTextBox.Visible = false;
+                    return;
                 }
+                previousAddrButton.Visible = true;
+                nextAddrButton.Visible = true;
+                urlTextBox.Visible = true;
+                qrBox.Visible = true;
+                UpdateQrCode(qrCodeText[qrCodeIndex]);
             }
         }
 
@@ -293,6 +445,11 @@ namespace CleverDecksLauncher
         private void checkBox1_CheckedChanged(object sender, EventArgs e)
         {
 
+        }
+
+        private void toolstripMenuItem_launchCD_Click(object sender, EventArgs e)
+        {
+            LaunchBrowserMenuItem_Click(sender, e);
         }
     }
 }
